Checking connection health via Proton URL
We we using a Mullvad service (https://am.i.mullvad.net/) to confirm that connection
was working and to retrieve some GeoIP data (City, ISP and IP Address).

Tests showed that in cases where connection was not working properly, Mullvad endpoint
was not responding while ip.me was.

Given ip.me is operated by ProtonVPN, this will allow us to live-test whether this
helps with our connection issues.